Timing Issue Fix
The timing issue arises from the (non-existent) gap between the deploy and register phases. The deploy errand does a restart of the app and does not wait before kicking off the register errand. On vsphere, the time to spin up a new vm to run the register is very small that the app is not fully up and register fails. On Aws, it takes longer and the app is up by then.

In order to avoid this surprise and hassle, I plan to add a bit of check-sleep-check for service broker app health before proceeding with the register.